+++ src/public.js
@@ -6,8 +6,8 @@
  *
  * @class Grid
  * @constructor
- * @param [element] {Object} The corresponding DOM element.
- * @param [options] {Object} The options to initialize the plugin.
+ * @param element {Object} The corresponding DOM element.
+ * @param options {Object} The options to override default settings.
  * @chainable
  **/
 var Grid = function(element, options)
@@ -17,12 +17,14 @@ var Grid = function(element, options)
     // overrides rowCount explicitly because deep copy ($.extend) leads to strange behaviour
     var rowCount = this.options.rowCount = this.element.data().rowCount || options.rowCount || this.options.rowCount;
     this.columns = [];
-    this.identifier = null; // The first column ID that is marked as identifier
     this.current = 1;
-    this.rows = [];
+    this.currentRows = [];
+    this.identifier = null; // The first column ID that is marked as identifier
     this.rowCount = ($.isArray(rowCount)) ? rowCount[0] : rowCount;
-    this.sort = {};
+    this.rows = [];
     this.searchPhrase = "";
+    this.selectedRows = [];
+    this.sort = {};
     this.total = 0;
     this.totalPages = 0;
     this.cachedParams = {
@@ -40,16 +42,16 @@ Grid.defaults = {
     navigation: 3, // it's a flag: 0 = none, 1 = top, 2 = bottom, 3 = both (top and bottom)
     padding: 2, // page padding (pagination)
     rowCount: [10, 25, 50, -1], // rows per page int or array of int (-1 represents "All")
-    selection: false, // todo: implement!
-    multiSelect: false, // todo: implement!
-    selectRows: false, // todo: implement and find a better name for this property [select new rows after adding]!
+    selection: false,
+    multiSelect: false,
+    highlightRows: false, // highlights new rows (find the page of the first new row)
     sorting: true,
     multiSort: false,
     ajax: false, // todo: find a better name for this property to differentiate between client-side and server-side data
     post: {}, // or use function () { return {}; } (reserved properties are "current", "rowCount", "sort" and "searchPhrase")
     url: "", // or use function () { return ""; }
 
-    // note: The following properties should be used via data-api attributes
+    // note: The following properties should not be used via data-api attributes
     converters: {
         numeric: {
             from: function (value) { return +value; }, // converts from string to numeric
@@ -66,6 +68,7 @@ Grid.defaults = {
         center: "text-center",
         columnHeaderAnchor: "column-header-anchor", // must be a unique class name or constellation of class names within the column header cell
         columnHeaderText: "text",
+        dropDownItem: "dropdown-item", // must be a unique class name or constellation of class names within the actionDropDown,
         dropDownItemButton: "dropdown-item-button", // must be a unique class name or constellation of class names within the actionDropDown
         dropDownItemCheckbox: "dropdown-item-checkbox", // must be a unique class name or constellation of class names within the actionDropDown
         dropDownMenu: "dropdown btn-group", // must be a unique class name or constellation of class names within the actionDropDown
@@ -102,8 +105,8 @@ Grid.defaults = {
     templates: {
         actionButton: "<button class=\"btn btn-default\" type=\"button\" title=\"{{ctx.text}}\">{{ctx.content}}</button>",
         actionDropDown: "<div class=\"{{css.dropDownMenu}}\"><button class=\"btn btn-default dropdown-toggle\" type=\"button\" data-toggle=\"dropdown\"><span class=\"{{css.dropDownMenuText}}\">{{ctx.content}}</span> <span class=\"caret\"></span></button><ul class=\"{{css.dropDownMenuItems}}\" role=\"menu\"></ul></div>",
-        actionDropDownItem: "<li><a href=\"{{ctx.uri}}\" class=\"{{css.dropDownItemButton}}\">{{ctx.text}}</a></li>",
-        actionDropDownCheckboxItem: "<li><label class=\"{{css.dropDownItemCheckbox}}\"><input name=\"{{ctx.name}}\" type=\"checkbox\" value=\"1\" {{ctx.checked}} /> {{ctx.label}}</label></li>",
+        actionDropDownItem: "<li><a href=\"{{ctx.uri}}\" class=\"{{css.dropDownItem}} {{css.dropDownItemButton}}\">{{ctx.text}}</a></li>",
+        actionDropDownCheckboxItem: "<li><label class=\"{{css.dropDownItem}}\"><input name=\"{{ctx.name}}\" type=\"checkbox\" value=\"1\" class=\"{{css.dropDownItemCheckbox}}\" {{ctx.checked}} /> {{ctx.label}}</label></li>",
         actions: "<div class=\"{{css.actions}}\"></div>",
         body: "<tbody></tbody>",
         cell: "<td class=\"{{ctx.css}}\">{{ctx.content}}</td>",
@@ -123,26 +126,46 @@ Grid.defaults = {
     }
 };
 
+/**
+ * Appends rows.
+ *
+ * @method append
+ * @param rows {Array} An array of rows to append
+ * @chainable
+ **/
 Grid.prototype.append = function(rows)
 {
-    // there is only support for client-side data
-    if (!this.options.ajax)
+    if (this.options.ajax)
+    {
+        // todo: implement ajax DELETE
+    }
+    else
     {
         for (var i = 0; i < rows.length; i++)
         {
             appendRow.call(this, rows[i]);
         }
         sortRows.call(this);
+        highlightAppendedRows.call(this, rows);
         loadData.call(this);
     }
 
     return this;
 };
 
+/**
+ * Removes all rows.
+ *
+ * @method clear
+ * @chainable
+ **/
 Grid.prototype.clear = function()
 {
-    // there is only support for client-side data
-    if (!this.options.ajax)
+    if (this.options.ajax)
+    {
+        // todo: implement ajax POST
+    }
+    else
     {
         this.rows = [];
         this.current = 1;
@@ -153,6 +176,12 @@ Grid.prototype.clear = function()
     return this;
 };
 
+/**
+ * Removes the control functionality completely and transforms the current state to the initial HTML structure.
+ *
+ * @method destroy
+ * @chainable
+ **/
 Grid.prototype.destroy = function()
 {
     // todo: this method has to be optimized (the complete initial state must be restored)
@@ -170,6 +199,12 @@ Grid.prototype.destroy = function()
     return this;
 };
 
+/**
+ * Resets the state and reloads rows.
+ *
+ * @method reload
+ * @chainable
+ **/
 Grid.prototype.reload = function()
 {
     this.current = 1; // reset
@@ -178,17 +213,57 @@ Grid.prototype.reload = function()
     return this;
 };
 
-Grid.prototype.remove = function(id)
+/**
+ * Removes rows by ids. Removes selected rows if no ids are provided.
+ *
+ * @method remove
+ * @param [rowsIds] {Array} An array of rows ids to remove
+ * @chainable
+ **/
+Grid.prototype.remove = function(rowIds)
 {
-    // there is only support for client-side data
-    if (!this.options.ajax) // check also is identifier available
+    if (this.identifier != null)
     {
-        // todo: implement!
+        var that = this;
+
+        if (this.options.ajax)
+        {
+            // todo: implement ajax DELETE
+        }
+        else
+        {
+            rowIds = rowIds || this.selectedRows;
+            var id;
+
+            for (var i = 0; i < rowIds.length; i++)
+            {
+                id = rowIds[i];
+
+                for (var j = 0; j < this.rows.length; j++)
+                {
+                    if (this.rows[j][this.identifier] === id)
+                    {
+                        this.rows.splice(j, 1);
+                        break;
+                    }
+                }
+            }
+
+            this.current = 1; // reset
+            loadData.call(this);
+        }
     }
 
     return this;
 };
 
+/**
+ * Searches in all rows for a specific phrase (but only in visible cells).
+ *
+ * @method search
+ * @param phrase {String} The phrase to search for
+ * @chainable
+ **/
 Grid.prototype.search = function(phrase)
 {
     if (this.searchPhrase !== phrase)
@@ -201,13 +276,13 @@ Grid.prototype.search = function(phrase)
     return this;
 };
 
-Grid.prototype.select = function()
-{
-    // todo: implement!
-
-    return this;
-};
-
+/**
+ * Sorts rows.
+ *
+ * @method sort
+ * @param dictionary {Object} A dictionary which contains the sort information
+ * @chainable
+ **/
 Grid.prototype.sort = function(dictionary)
 {
     var values = (dictionary) ? $.extend({}, dictionary) : {};